updates mui structure and adds palette definition
