+++ lib/mui/themes/main/palette.ts
@@ -1,0 +1,10 @@
+const palette = {
+  green: {
+    main: "#00B712",
+  },
+  neutral: {
+    main: "#C4C4C4",
+  },
+};
+
+export default palette;
