--- lib/themes/main/palette.ts
@@ -1,17 +1,0 @@
-declare module "@mui/material/SvgIcon" {
-  interface SvgIconPropsColorOverrides {
-    green: true;
-    grey: true;
-  }
-}
-
-const palette = {
-  green: {
-    main: "#00B712",
-  },
-  grey: {
-    main: "#C4C4C4",
-  },
-};
-
-export default palette;
